requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
beautifulsoup4==4.14.3
boto3==1.42.58
botocore==1.42.58
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dill==0.4.1
distro==1.9.0
doclayout_yolo==0.0.2b1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
loguru==0.7.3
magic-pdf==1.3.12
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.34.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==7.34.0
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.14
pyparsing==3.3.2
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rapid-table==1.0.5
rdkit==2025.9.6
regex==2026.2.19
requests==2.32.5
robust-downloader==0.0.2
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
simsimd==6.5.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stringzilla==4.6.0
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.18
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
win32_setctime==1.2.0
xlsxwriter==3.2.9
